Adds game
