+++ PingPong/PlayersNameSelector.hpp
@@ -5,21 +5,7 @@ TextBox pl1("Player 1 name", &arial, 20, 200, 50, 10, 10, sf::Color::Red, sf::Co
 TextInput Player2("", &arial, 16, 200, 50, 590, 70, sf::Color::White, sf::Color::Black);
 TextBox pl2("Player 2 name", &arial, 20, 200, 50, 590, 10, sf::Color::Red, sf::Color::Black);
 Button per("Submit", &arial, 20, 100, 50, 350, 10, sf::Color::Green, sf::Color::Black);
-PhysicalBody ball(50, 50, 10, 10);
-void Gamest() {
-	nameSelector.disable();
-	game.enable();
-	app.addScreen(&game);
-	game.addObject(&ball);
-	ball.setTexture(&balltexture);
-	ball.setXSpeed(1);
-	ball.setYSpeed(1);
-	ball.setColideBorder(true);
-	ball.process();
-	//ball.move(0.1,0.1);
 
-
-}
 void callback() {
 	if (mode != singlePlayer) {
 		if (Player1.getLine() == "") {
