+++ PingPong/main.cpp
@@ -2,6 +2,7 @@
 #include "GlobalDeclaration.hpp"
 #include "MainMenu.hpp";
 #include "PlayersNameSelector.hpp"
+#include "GameRun.hpp"
